+++ packages/iov-crypto/src/englishmnemonic.spec.ts
@@ -1,10 +1,8 @@
-import { Encoding, fromBase64, fromHex } from "@iov/encoding";
+import { fromAscii, fromBase64, fromHex } from "@iov/encoding";
 
 import { EnglishMnemonic } from "./englishmnemonic";
 import { Sha256 } from "./sha";
 import wordlists from "./testdata/bip39_wordlists.json";
-
-const { fromAscii } = Encoding;
 
 describe("EnglishMnemonic", () => {
   describe("wordlist", () => {