+++ packages/iov-crypto/src/libsodium.spec.ts
@@ -1,5 +1,5 @@
 /* tslint:disable:no-bitwise */
-import { Encoding, fromHex } from "@iov/encoding";
+import { fromHex, toAscii } from "@iov/encoding";
 
 import {
   Argon2id,
@@ -12,8 +12,6 @@ import {
   Xchacha20poly1305IetfMessage,
   Xchacha20poly1305IetfNonce,
 } from "./libsodium";
-
-const { toAscii } = Encoding;
 
 describe("Libsodium", () => {
   describe("Argon2id", () => {