+++ packages/iov-encoding/src/encoding.spec.ts
@@ -1,31 +1,6 @@
 import { Encoding } from "./encoding";
 
 describe("Encoding", () => {
-  it("encodes to ascii", () => {
-    expect(Encoding.toAscii("")).toEqual(new Uint8Array([]));
-    expect(Encoding.toAscii("abc")).toEqual(new Uint8Array([0x61, 0x62, 0x63]));
-    expect(Encoding.toAscii(" ?=-n|~+-*/\\")).toEqual(
-      new Uint8Array([0x20, 0x3f, 0x3d, 0x2d, 0x6e, 0x7c, 0x7e, 0x2b, 0x2d, 0x2a, 0x2f, 0x5c]),
-    );
-
-    expect(() => Encoding.toAscii("ö")).toThrow();
-    expect(() => Encoding.toAscii("ß")).toThrow();
-  });
-
-  it("decodes from ascii", () => {
-    expect(Encoding.fromAscii(new Uint8Array([]))).toEqual("");
-    expect(Encoding.fromAscii(new Uint8Array([0x61, 0x62, 0x63]))).toEqual("abc");
-    expect(
-      Encoding.fromAscii(
-        new Uint8Array([0x20, 0x3f, 0x3d, 0x2d, 0x6e, 0x7c, 0x7e, 0x2b, 0x2d, 0x2a, 0x2f, 0x5c]),
-      ),
-    ).toEqual(" ?=-n|~+-*/\\");
-
-    expect(() => Encoding.fromAscii(new Uint8Array([0x00]))).toThrow();
-    expect(() => Encoding.fromAscii(new Uint8Array([0x7f]))).toThrow();
-    expect(() => Encoding.fromAscii(new Uint8Array([0xff]))).toThrow();
-  });
-
   describe("UTF8 encoding", () => {
     it("encodes ascii strings", () => {
       expect(Encoding.toUtf8("")).toEqual(new Uint8Array([]));