+++ packages/iov-keycontrol/src/wallets/secp256k1hdwallet.spec.ts
@@ -1,10 +1,8 @@
 import { ChainId, PrehashType, SignableBytes } from "@iov/bcp";
 import { ExtendedSecp256k1Signature, Secp256k1, Sha256, Slip10RawIndex } from "@iov/crypto";
-import { Encoding, fromHex } from "@iov/encoding";
+import { fromHex, toAscii } from "@iov/encoding";
 
 import { Secp256k1HdWallet } from "./secp256k1hdwallet";
-
-const { toAscii } = Encoding;
 
 describe("Secp256k1HdWallet", () => {
   const defaultChain = "chain123" as ChainId;