+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## HEAD
 
-- @iov/encoding: Pull out free functions `fromHex`, `toHex`, `fromBase64` and
-  `toBase64` to simplify usability. `Encoding.fromHex`, `Encoding.toHex`,
-  `Encoding.fromBase64` and `Encoding.toBase64` are now deprecated.
+- @iov/encoding: Pull out free functions `fromAscii`, `toAscii`, `fromHex`,
+  `toHex`, `fromBase64` and `toBase64` to simplify usability.
+  `Encoding.fromAscii`, `Encoding.toAscii`, `Encoding.fromHex`,
+  `Encoding.toHex`, `Encoding.fromBase64` and `Encoding.toBase64` are now
+  deprecated.
 
 ## 2.2.5